helper.js renamed to bootstrapElements
