+++ index.html
@@ -37,13 +37,22 @@
             </button>
           </div>
           <div class="col-2 col-lg-1">
-            <div class="btn-group" role="group" aria-label="Sort and Filter">
-              <button type="button" class="btn btn-outline-secondary btn-sm">
-                Sort
+            <div class="dropdown">
+              <button
+                class="btn btn-secondary dropdown-toggle"
+                type="button"
+                data-bs-toggle="dropdown"
+                aria-expanded="false"
+              >
+                Dropdown button
               </button>
-              <button type="button" class="btn btn-outline-secondary btn-sm">
-                Filter
-              </button>
+              <ul class="dropdown-menu">
+                <li><a class="dropdown-item" href="#">Action</a></li>
+                <li><a class="dropdown-item" href="#">Another action</a></li>
+                <li>
+                  <a class="dropdown-item" href="#">Something else here</a>
+                </li>
+              </ul>
             </div>
           </div>
         </div>
